+++ changelog.md
@@ -7,6 +7,8 @@ Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 
 ## [Unreleased]
 
+### Added
+- Added the new method 'CollectionWorker.get_all_cursor'.
 
 ### Changed
 - Improved the annotation of motor objects.